Update daily_RACMO-to-zarr.ipynb
correct bug with time vector caused by not decoding during loading of the netcdfs. I was not decoding because it was solving a later error with writing to zarr, but I didnt realize it was causing the time vector to be completely wrong. So I had to instead solve the issue the writing to zarr. I did that by making sure the encoding of the datarray was self consistent. Specifically that missing_value and _FillValue are the same.